requirements.txt
djangorestframework==3.15.2
Django==5.1.2
pre_commit==4.0.1
